move common parts from bert_question_answering_demo and bert_question_answerint_embedding_demo into common as discussed
